+++ backend/internal/services/table.go
@@ -25,6 +25,11 @@ func (s *TableService) GetAll() ([]string, error) {
 		return []string{}, fmt.Errorf("Failed to get driver and connection")
 	}
 
+	if s.Connection.GetDatabase() == "" {
+		logger.Error("No database selected.")
+		return []string{}, fmt.Errorf("No database selected.")
+	}
+
 	tables, err := driver.GetAllTable(conn.DB)
 	if err != nil {
 		logger.Error("Failed to list tables: %v", err)
@@ -39,6 +44,11 @@ func (s *TableService) Create(req models.Table) error {
 	if err != nil {
 		logger.Error("Failed to get driver and connection: %v", err)
 		return err
+	}
+
+	if s.Connection.GetDatabase() == "" {
+		logger.Error("No database selected.")
+		return fmt.Errorf("No database selected.")
 	}
 
 	err = driver.CreateTable(conn.DB, req)
@@ -57,6 +67,11 @@ func (s *TableService) Update(oldName, newName string) error {
 		return err
 	}
 
+	if s.Connection.GetDatabase() == "" {
+		logger.Error("No database selected.")
+		return fmt.Errorf("No database selected.")
+	}
+
 	err = driver.UpdateTable(conn.DB, oldName, newName)
 	if err != nil {
 		logger.Error("Failed to update table: %v", err)
@@ -71,6 +86,11 @@ func (s *TableService) Delete(table string) error {
 	if err != nil {
 		logger.Error("Failed to get driver and connection: %v", err)
 		return err
+	}
+
+	if s.Connection.GetDatabase() == "" {
+		logger.Error("No database selected.")
+		return fmt.Errorf("No database selected.")
 	}
 
 	err = driver.DeleteTable(conn.DB, table)
